[AMD] Replace GLM-5.1 MI35x nightly with GLM-5.2 perf
Add a cookbook-aligned performance sweep after GLM-5.2 accuracy and retire the redundant standard GLM-5.1 MI35x jobs, preserving MI30x and specialized GLM-5.1 coverage while reducing scarce runner usage.
