--- test/registered/amd/accuracy/mi35x/test_glm51_eval_mi35x.py
@@ -1,238 +1,0 @@
-"""MI35x GLM-5.1 GSM8K Completion Evaluation Test (8-GPU)
-
-Tests GLM-5.1-FP8 with DSA attention backend using few-shot
-completion benchmark on MI35x.
-
-Registry: nightly-amd-8-gpu-mi35x-glm51 suite
-"""
-
-import ast
-import os
-import re
-import time
-import unittest
-from dataclasses import dataclass, field
-from typing import List, Optional, Tuple
-
-import numpy as np
-
-from sglang.srt.utils import kill_process_tree
-from sglang.test.ci.ci_register import register_amd_ci
-from sglang.test.test_utils import (
-    DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
-    DEFAULT_URL_FOR_TEST,
-    is_in_ci,
-    popen_launch_server,
-    write_github_step_summary,
-)
-from sglang.utils import download_and_cache_file, read_jsonl
-
-register_amd_ci(
-    est_time=5400,
-    suite="nightly-amd-8-gpu-mi35x-glm51",
-    nightly=True,
-)
-
-INVALID = -9999999
-
-
-@dataclass
-class ModelConfig:
-    model_path: str
-    tp_size: int = 8
-    accuracy_threshold: float = 0.50
-    other_args: List[str] = field(default_factory=list)
-    env_vars: dict = field(default_factory=dict)
-    timeout: Optional[int] = None
-    variant: Optional[str] = None
-
-    def get_display_name(self) -> str:
-        if self.variant:
-            return f"{self.model_path} ({self.variant})"
-        return self.model_path
-
-
-MI35X_GLM51_MODELS = [
-    ModelConfig(
-        model_path="zai-org/GLM-5.1-FP8",
-        tp_size=8,
-        accuracy_threshold=0.93,
-        timeout=5400,
-        variant="dsa",
-        other_args=[
-            "--trust-remote-code",
-            "--reasoning-parser",
-            "glm45",
-            "--tool-call-parser",
-            "glm47",
-            "--dsa-prefill-backend",
-            "tilelang",
-            "--dsa-decode-backend",
-            "tilelang",
-            "--chunked-prefill-size",
-            "131072",
-            "--mem-fraction-static",
-            "0.80",
-            "--model-loader-extra-config",
-            '{"enable_multithread_load": true}',
-            "--watchdog-timeout",
-            "1200",
-        ],
-        env_vars={},
-    ),
-]
-
-
-def get_one_example(lines, i, include_answer):
-    ret = "Question: " + lines[i]["question"] + "\nAnswer:"
-    if include_answer:
-        ret += " " + lines[i]["answer"]
-    return ret
-
-
-def get_few_shot_examples(lines, k):
-    ret = ""
-    for i in range(k):
-        ret += get_one_example(lines, i, True) + "\n\n"
-    return ret
-
-
-def get_answer_value(answer_str):
-    answer_str = answer_str.replace(",", "")
-    numbers = re.findall(r"\d+", answer_str)
-    if len(numbers) < 1:
-        return INVALID
-    try:
-        return ast.literal_eval(numbers[-1])
-    except SyntaxError:
-        return INVALID
-
-
-def run_gsm8k_benchmark(
-    base_url: str,
-    num_questions: int = 200,
-    num_shots: int = 5,
-    parallel: int = 64,
-) -> Tuple[float, float, float]:
-    import sglang as sgl
-    from sglang.lang.backend.runtime_endpoint import RuntimeEndpoint
-
-    url = "https://raw.githubusercontent.com/openai/grade-school-math/master/grade_school_math/data/test.jsonl"
-    data_path = download_and_cache_file(url)
-    lines = list(read_jsonl(data_path))
-
-    few_shot_examples = get_few_shot_examples(lines, num_shots)
-
-    questions = []
-    labels = []
-    for i in range(len(lines[:num_questions])):
-        questions.append(get_one_example(lines, i, False))
-        labels.append(get_answer_value(lines[i]["answer"]))
-    assert all(l != INVALID for l in labels)
-    arguments = [{"question": q} for q in questions]
-
-    @sgl.function
-    def few_shot_gsm8k(s, question):
-        s += few_shot_examples + question
-        s += sgl.gen(
-            "answer", max_tokens=512, stop=["Question", "Assistant:", "<|separator|>"]
-        )
-
-    backend = RuntimeEndpoint(base_url)
-    sgl.set_default_backend(backend)
-
-    tic = time.perf_counter()
-    states = few_shot_gsm8k.run_batch(
-        arguments, temperature=0, num_threads=parallel, progress_bar=True
-    )
-    latency = time.perf_counter() - tic
-
-    preds = [get_answer_value(states[i]["answer"]) for i in range(len(states))]
-    acc = np.mean(np.array(preds) == np.array(labels))
-    invalid = np.mean(np.array(preds) == INVALID)
-
-    return float(acc), float(invalid), float(latency)
-
-
-class TestGLM51EvalMI35x(unittest.TestCase):
-    """GLM-5.1 GSM8K Completion Evaluation Test for AMD MI35x."""
-
-    @classmethod
-    def setUpClass(cls):
-        cls.models = MI35X_GLM51_MODELS
-        cls.base_url = DEFAULT_URL_FOR_TEST
-        cls.num_questions = int(os.environ.get("GSM8K_NUM_QUESTIONS", "1319"))
-
-    def test_glm51_accuracy(self):
-        all_results = []
-        summary = "### GLM-5.1 Models (MI35x)\n\n"
-        summary += "| Model | Variant | TP | Accuracy | Threshold | Status |\n"
-        summary += "| ----- | ------- | -- | -------- | --------- | ------ |\n"
-
-        for config in self.models:
-            display_name = config.get_display_name()
-            with self.subTest(model=display_name):
-                print(f"\n{'='*60}")
-                print(f"Testing: {display_name}")
-                print(f"{'='*60}")
-
-                env = os.environ.copy()
-                for key, value in config.env_vars.items():
-                    env[key] = value
-
-                other_args = list(config.other_args)
-                other_args.extend(["--tp", str(config.tp_size)])
-                timeout = config.timeout or DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH
-
-                try:
-                    process = popen_launch_server(
-                        model=config.model_path,
-                        base_url=self.base_url,
-                        timeout=timeout,
-                        other_args=other_args,
-                        env=env,
-                    )
-
-                    try:
-                        acc, invalid, latency = run_gsm8k_benchmark(
-                            self.base_url, num_questions=self.num_questions
-                        )
-                        passed = acc >= config.accuracy_threshold
-                        status = "✅ PASS" if passed else "❌ FAIL"
-                        print(
-                            f"  accuracy={acc:.3f} threshold={config.accuracy_threshold} {status}"
-                        )
-
-                        all_results.append(
-                            {
-                                "model": display_name,
-                                "accuracy": acc,
-                                "passed": passed,
-                            }
-                        )
-                        summary += f"| {config.model_path} | {config.variant or 'N/A'} | {config.tp_size} | {acc:.3f} | {config.accuracy_threshold} | {status} |\n"
-
-                    finally:
-                        kill_process_tree(process.pid)
-
-                except Exception as e:
-                    summary += f"| {config.model_path} | {config.variant or 'N/A'} | {config.tp_size} | N/A | {config.accuracy_threshold} | ❌ ERROR |\n"
-                    all_results.append(
-                        {
-                            "model": display_name,
-                            "accuracy": None,
-                            "passed": False,
-                            "error": str(e),
-                        }
-                    )
-
-        if is_in_ci():
-            write_github_step_summary(summary)
-
-        failed = [r for r in all_results if not r["passed"]]
-        if failed:
-            raise AssertionError(f"Failed models: {[r['model'] for r in failed]}")
-
-
-if __name__ == "__main__":
-    unittest.main()
