--- test/registered/amd/perf/mi35x/test_glm51_perf_mi35x.py
@@ -1,141 +1,0 @@
-"""MI35x Nightly performance benchmark for GLM-5.1.
-
-Tests GLM-5.1-FP8 with DSA attention backend using bench_one_batch
-on 8 GPUs with TP=8, FP8 KV cache.
-
-Registry: nightly-perf-8-gpu-mi35x-glm51 suite
-"""
-
-import os
-import unittest
-from typing import List
-
-from sglang.test.ci.ci_register import register_amd_ci
-from sglang.test.nightly_bench_utils import BenchmarkResult
-from sglang.test.nightly_utils import NightlyBenchmarkRunner
-from sglang.test.test_utils import DEFAULT_URL_FOR_TEST, _parse_int_list_env
-
-register_amd_ci(est_time=5400, suite="nightly-perf-8-gpu-mi35x-glm51", nightly=True)
-
-
-def generate_simple_markdown_report(results: List[BenchmarkResult]) -> str:
-    model_header = results[0].model_path
-    if results[0].run_name and results[0].run_name != "default":
-        model_header += f" ({results[0].run_name})"
-
-    gpu_config = os.getenv("GPU_CONFIG", "MI35x")
-    if gpu_config:
-        model_header += f" [{gpu_config}]"
-
-    summary = f"### {model_header}\n"
-    summary += "| batch size | input len | latency (s) | input throughput (tok/s) | output throughput (tok/s) | ITL (ms) |\n"
-    summary += "| ---------- | --------- | ----------- | ------------------------ | ------------------------- | -------- |\n"
-
-    report_results = (
-        results[1:]
-        if len(results) > 1 and results[0].batch_size == results[1].batch_size
-        else results
-    )
-
-    for result in report_results:
-        itl = 1 / (result.output_throughput / result.batch_size) * 1000
-        summary += f"| {result.batch_size} | {result.input_len} | {result.latency:.2f} | {result.input_throughput:.2f} | {result.output_throughput:.2f} | {itl:.2f} |\n"
-
-    return summary
-
-
-GLM51_MODEL_PATH = os.environ.get("GLM51_MODEL_PATH", "zai-org/GLM-5.1-FP8")
-RESULT_DIR = "performance_results_glm51_mi35x"
-
-
-class TestGLM51PerfMI35x(unittest.TestCase):
-    """Nightly performance benchmark for GLM-5.1 on MI35x.
-
-    Tests GLM-5.1-FP8 with DSA attention backend on TP=8.
-    """
-
-    @classmethod
-    def setUpClass(cls):
-        cls.base_url = DEFAULT_URL_FOR_TEST
-        cls.batch_sizes = [1, 8, 16, 64]
-        cls.input_lens = tuple(_parse_int_list_env("NIGHTLY_INPUT_LENS", "4096"))
-        cls.output_lens = tuple(_parse_int_list_env("NIGHTLY_OUTPUT_LENS", "512"))
-
-        cls.model_config = {
-            "name": "glm51-mi35x",
-            "model_path": GLM51_MODEL_PATH,
-            "other_args": [
-                "--trust-remote-code",
-                "--reasoning-parser",
-                "glm45",
-                "--tool-call-parser",
-                "glm47",
-                "--tp",
-                "8",
-                "--dsa-prefill-backend",
-                "tilelang",
-                "--dsa-decode-backend",
-                "tilelang",
-                "--kv-cache-dtype",
-                "fp8_e4m3",
-                "--chunked-prefill-size",
-                "131072",
-                "--mem-fraction-static",
-                "0.85",
-                "--model-loader-extra-config",
-                '{"enable_multithread_load": true, "num_threads": 8}',
-                "--watchdog-timeout",
-                "1200",
-            ],
-            "env_vars": {
-                "SGLANG_USE_AITER": "1",
-                "SGLANG_ROCM_FUSED_DECODE_MLA": "0",
-                "ROCM_QUICK_REDUCE_QUANTIZATION": "INT4",
-                "SAFETENSORS_FAST_GPU": "1",
-            },
-        }
-
-        os.environ.setdefault("SGLANG_BENCH_TIMEOUT", "3600")
-        cls.runner = NightlyBenchmarkRunner(RESULT_DIR, cls.__name__, cls.base_url)
-        cls.runner.setup_result_directory()
-        cls.runner.full_report = f"## {cls.__name__}\n"
-
-    def test_glm51_perf(self):
-        old_env = {}
-        for key, value in self.model_config.get("env_vars", {}).items():
-            old_env[key] = os.environ.get(key)
-            os.environ[key] = value
-
-        try:
-            result_tuple = self.runner.run_benchmark_for_model(
-                model_path=self.model_config["model_path"],
-                batch_sizes=self.batch_sizes,
-                input_lens=self.input_lens,
-                output_lens=self.output_lens,
-                other_args=self.model_config["other_args"],
-                variant=self.model_config["name"],
-                extra_bench_args=["--trust-remote-code"],
-                timeout=5400,
-            )
-            results = result_tuple[0]
-            success = result_tuple[1]
-
-            if results:
-                self.runner.full_report += (
-                    generate_simple_markdown_report(results) + "\n"
-                )
-
-            self.assertTrue(
-                success, f"Benchmark failed for {GLM51_MODEL_PATH} on MI35x"
-            )
-        finally:
-            for key, value in old_env.items():
-                if value is None:
-                    os.environ.pop(key, None)
-                else:
-                    os.environ[key] = value
-            self.runner.write_final_report()
-
-
-if __name__ == "__main__":
-    unittest.main()
