+++ test/run_suite.py
@@ -142,6 +142,7 @@ NIGHTLY_SUITES = {
         "nightly-amd-8-gpu-mi35x-deepseek-r1-mxfp4-tp4",
         "nightly-amd-accuracy-8-gpu-mi35x-kimi-k3",
         "nightly-amd-8-gpu-mi35x-glm52-fp8",
+        "nightly-perf-8-gpu-mi35x-glm52-fp8",
         "nightly-amd-4-gpu",
         "nightly-amd-8-gpu",
         "nightly-amd-vlm",
